Add API formatTableWithContentModel
