+++ packages-content-model/roosterjs-content-model-api/lib/index.ts
@@ -40,4 +40,6 @@ export { default as adjustImageSelection } from './publicApi/image/adjustImageSe
 export { default as setParagraphMargin } from './publicApi/block/setParagraphMargin';
 export { default as toggleCode } from './publicApi/segment/toggleCode';
 export { default as insertEntity } from './publicApi/entity/insertEntity';
+
+export { formatTableWithContentModel } from './publicApi/utils/formatTableWithContentModel';
 export { setListType } from './modelApi/list/setListType';
